theme updates. adjusted typography and spacing on home and problem details pages. increased line heights to improve readability.
