+++ src/components/problems/problem-detail-navbar/HintByDatabase.js
@@ -5,14 +5,34 @@
  * It can also be a video link (also provided by the icpc team).
  */
 
-import React from 'react'
-import NavbarStack from './NavbarStack'
+import React, { lazy } from 'react'
+import Typography from '@mui/material/Typography'
+import { styled } from '@mui/material/styles'
+import Box from '@mui/material/Box'
 
+// Dynamically import NavbarStack to optimize bundle size
+const NavbarStack = lazy(() => import('./NavbarStack'))
+
+const StyledTypography = styled(Typography)(({ theme }) => ({
+  marginBottom: theme.spacing(1),
+  marginTop: theme.spacing(1),
+}))
 const HintByDatabaseTab = ({ problem }) => {
   return (
     <NavbarStack>
-      <h3>Hint By Database</h3>
-      <p>{problem.hint}</p>
+      <Box>
+        <StyledTypography variant="h6" sx={{ fontWeight: 'bold' }}>
+          Hint By Database
+        </StyledTypography>
+        <Typography
+          variant="small"
+          sx={{
+            lineHeight: '1.75rem',
+          }}
+        >
+          {problem.hint}
+        </Typography>
+      </Box>
     </NavbarStack>
   )
 }
